More Panels and Thoughts
Added more panels that relate to the Weather API that Alexia used. My Last.fm API idea hasn't worked so far in respect to pulling audio decibel levels and setting a range to the data I got. If anyone knows of a way to do this, let me know. I will most likely add a range to the amount of plays a song gets and call it good.